Iterative filtering and pruning
When number of tables becomes large, it is not feasible to evaluate all
possible rel combinations in one shot (scratchpad size would become
too large).
Therefore, there is now a maximum scratchpad size (1GB). If the number
of combinations exceeds this number, the enumeration is split in
as many steps as it is necessary. After each enumeration step, filtering
is run on the new combinations in order to compact valid pairs on one
side. When the number of combination is exhausted or the number of
combinations to evaluate (after filtering) exceeds a certain threshold
(half scratchpad size), pruning is started as before. If there are still
some combinations to evaluate (very unlikely if number of tables is
reasonable), the pruned combinations in the memo are reloaded to
the scratchpad and everything continues as before (NB this may be
inefficient, especially since cost functions are evaluated twice).