Renamed enumerate to unrank
"A ranking algorithm determines the position (or rank) of a combinatorial
object among all the objects (with respect to a given order); an unranking
algorithm finds the object having a specified rank. Thus, ranking and
unranking can be considered as inverse operations."
It is clear that what we called "enumerate" is indeed an unranking
operation.

In the following, enumerate will be used for the different ways we can
list the candidate join pairs (i.e. DPsize, DPsub, DPccp)