Fix bucket selection in DependencyBuffer
join_rel->left_relation_id is the id of the left relation of the
currently best join relation, which usually differs from left_rel->id
which is the left relation of the relation pair to be joined.
This bug created race conditions since dependencies among join pairs
were not respected.