hastable: add self-resizing up to max dimension
hashtable will double itself when the number of added elements exceeds
half of its current capacity

NB: in general, the number of added elements is an upper bound of the
number of stored elements.
Since dpsub will not store duplicates and will not delete elements
(not even implemented), they are the same.