hashtable: fix resize loop problem
since resize used the same insert api as the user, the elements it moved
were counted into the number of elements inside the hashtable
therefore causing a nested resize until max size was reached.
The problem is solved by introducing a private _insert function that
only does the inserting and is called by resize and insert.